Adjust default dtype to match others
This previously never threw an error due to dtypes not being handled
correctly. With the dtype propagation fix in 5e904ef, we'd either have
to

- pass the `torch.float32` dtype explicitly everywhere or
- adjust the default like done here.

+++ pytorch3d/transforms/transform3d.py
@@ -598,7 +598,7 @@ class RotateAxisAngle(Rotate):
         angle,
         axis: str = "X",
         degrees: bool = True,
-        dtype: torch.dtype = torch.float64,
+        dtype: torch.dtype = torch.float32,
         device: Optional[Device] = None,
     ) -> None:
         """
